# Make Gradle tasks cacheable