--- a/jte-gradle-plugin/build.gradle
+++ b/jte-gradle-plugin/build.gradle
@@ -6,8 +6,8 @@ plugins {
 }
 
 repositories {
-    mavenCentral()
     mavenLocal()
+    mavenCentral()
 }
 
 dependencies {