--- a/jte-gradle-plugin/src/main/java/gg/jte/gradle/GenerateJteTask.java
+++ b/jte-gradle-plugin/src/main/java/gg/jte/gradle/GenerateJteTask.java
@@ -1,6 +1,7 @@
 package gg.jte.gradle;
 
 import org.gradle.api.file.ConfigurableFileCollection;
+import org.gradle.api.tasks.CacheableTask;
 import org.gradle.api.tasks.Classpath;
 import org.gradle.api.tasks.InputFiles;
 import org.gradle.api.tasks.TaskAction;
@@ -10,6 +11,7 @@ import org.gradle.workers.WorkerExecutor;
 import javax.inject.Inject;
 import java.nio.file.Path;
 
+@CacheableTask
 public abstract class GenerateJteTask extends JteTaskBase {
 
     private final WorkerExecutor workerExecutor;