--- a/test/gradle-test-wrapper/build.gradle
+++ b/test/gradle-test-wrapper/build.gradle
@@ -1,6 +1,5 @@
 plugins {
     id('java')
-    id("com.adarshr.test-logger") version "4.0.0"
 }
 
 repositories {
@@ -17,8 +16,4 @@ dependencies {
 tasks.withType(Test).configureEach {
     useJUnitPlatform()
     systemProperty("gradle.matrix.versions", System.getProperty("gradle.matrix.versions", "DEFAULT"))
-}
-
-testlogger {
-    theme 'standard'
 }