--- a/test/jte-runtime-cp-test-models-gradle/settings.gradle
+++ b/test/jte-runtime-cp-test-models-gradle/settings.gradle
@@ -5,3 +5,9 @@ pluginManagement {
         gradlePluginPortal()
     }
 }
+
+buildCache {
+    local {
+        directory "${System.getProperty("test.build.cache.dir")}"
+    }
+}