--- a/test/jte-runtime-test-gradle-convention/settings.gradle
+++ b/test/jte-runtime-test-gradle-convention/settings.gradle
@@ -4,3 +4,9 @@ pluginManagement {
         gradlePluginPortal()
     }
 }
+
+buildCache {
+    local {
+        directory "${System.getProperty("test.build.cache.dir")}"
+    }
+}