## What?

Gradle has a build cache that can be used to speed up the build process. From [Gradle docs](https://docs.gradle.org/current/userguide/build_cache.html):

> The Gradle build cache is a cache mechanism that aims to save time by reusing outputs produced by other builds. The build cache works by storing (locally or remotely) build outputs and allowing builds to fetch these outputs from the cache when it is determined that inputs have not changed, avoiding the expensive work of regenerating them.

I noticed that the build cache for the `compileKotlin` task was not working when using jte's `generate` or `precompile`, because the inputs generated by those tasks were not cached. Therefore, the generated code would be considered new input for `compileKotlin`, meaning the task cannot use the build cache. I suspect the same applies to `compileJava`, but I didn't check.

## Local validation

Comparison when running the `compileKotlin` task with and without `-Dorg.gradle.caching=true`.

### Before

Running with `--build-cache` multiple times has no effect since the cache is always stale:

```
time -h -p gradle --warning-mode=none --build-cache clean compileKotlin
> Task :cleanGenerateJte
> Task :cleanPrecompileJte UP-TO-DATE
> Task :clean
> Task :checkKotlinGradlePluginConfigurationErrors

> Task :generateJte
Extension gg.jte.models.generator.ModelExtension generated 3 files.

> Task :kspKotlin
> Task :compileKotlin

BUILD SUCCESSFUL in 6s
7 actionable tasks: 6 executed, 1 up-to-date
real 6.96
user 1.10
sys 0.16
```

### After

With no cache (`--warning-mode=none` for brevity):

```
time -h -p gradle --warning-mode=none clean compileKotlin
> Task :cleanGenerateJte
> Task :cleanPrecompileJte UP-TO-DATE
> Task :clean
> Task :checkKotlinGradlePluginConfigurationErrors

> Task :generateJte
Extension gg.jte.models.generator.ModelExtension generated 3 files.

> Task :kspKotlin
> Task :compileKotlin

BUILD SUCCESSFUL in 8s
7 actionable tasks: 6 executed, 1 up-to-date
real 9.30
user 1.17
sys 0.15
```

With a populated build-cache (notice the `FROM-CACHE`):

```
time -h -p gradle --warning-mode=none --build-cache clean compileKotlin
> Task :cleanGenerateJte
> Task :cleanPrecompileJte UP-TO-DATE
> Task :clean
> Task :checkKotlinGradlePluginConfigurationErrors
> Task :generateJte FROM-CACHE
> Task :kspKotlin FROM-CACHE
> Task :compileKotlin FROM-CACHE

BUILD SUCCESSFUL in 1s
7 actionable tasks: 3 executed, 3 from cache, 1 up-to-date
real 1.49
user 1.13
sys 0.16
```

Let's see the effect on `testClasses` after running the command above:


```
time -h -p gradle --warning-mode=none --build-cache clean testClasses
> Task :cleanGenerateJte
> Task :cleanPrecompileJte UP-TO-DATE
> Task :clean
> Task :checkKotlinGradlePluginConfigurationErrors
> Task :generateJte FROM-CACHE
> Task :kspKotlin FROM-CACHE
> Task :compileKotlin FROM-CACHE
> Task :compileJava NO-SOURCE
> Task :processResources
> Task :classes
> Task :kspTestKotlin
> Task :processTestResources
> Task :compileTestKotlin
> Task :compileTestJava NO-SOURCE
> Task :testClasses

BUILD SUCCESSFUL in 8s
11 actionable tasks: 7 executed, 3 from cache, 1 up-to-date
real 8.94
user 1.21
sys 0.18
```

And then, since the build cache was populated:

```
time -h -p gradle --warning-mode=none --build-cache clean testClasses
> Task :cleanGenerateJte
> Task :cleanPrecompileJte UP-TO-DATE
> Task :clean
> Task :checkKotlinGradlePluginConfigurationErrors
> Task :generateJte FROM-CACHE
> Task :kspKotlin FROM-CACHE
> Task :compileKotlin FROM-CACHE
> Task :compileJava NO-SOURCE
> Task :processResources
> Task :classes
> Task :kspTestKotlin FROM-CACHE
> Task :compileTestKotlin FROM-CACHE
> Task :compileTestJava NO-SOURCE
> Task :processTestResources
> Task :testClasses

BUILD SUCCESSFUL in 1s
11 actionable tasks: 5 executed, 5 from cache, 1 up-to-date
real 2.40
user 1.26
sys 0.16
```