@marcospereira 2023-11-03
@casid, I'm trying to figure out how to add a test for this, though.