@marcospereira 2023-11-03
> I don't have time to have a detailed look, but I skimmed through and it seems good.

Yeah, I'm trying to understand why the tests Work On My Machine™ but not on CI.