@marcospereira 2023-11-03
Hum, it seems things are... broken(?) somehow. From this run: https://github.com/casid/jte/actions/runs/6748107408/job/18345752283?pr=295#step:6:479


```
> Task :publishToMavenLocal
Skipping task ':publishToMavenLocal' as it has no actions.
```

🤔