@casid 2023-11-05
Unfortunately, I'm not really a gradle expert, so I can't help much with this. :-(