@marcospereira 2023-11-08
> Unfortunately, I'm not really a gradle expert, so I can't help much with this. :-(

No worries. I was able to figure out how to properly test the tasks when the build-cache is involved. 

I think this is good to review/merge. 🚀