@marcospereira 2023-11-14
> I am happy with this change. Thanks @marcospereira !

Awesome! Thanks for taking the time to review, @edward3h.

@casid, I guess this is good to merge/release then. 🍾